+++ xknx/io/knxip_interface.py
@@ -431,15 +431,15 @@ class KNXIPInterface:
             await self._interface.disconnect()
             self._interface = None
 
-    async def telegram_received(self, telegram: Telegram) -> list[Telegram] | None:
+    def telegram_received(self, telegram: Telegram) -> None:
         """Put received telegram into queue. Callback for having received telegram."""
         if isinstance(telegram.tpci, TDataGroup):
             self.xknx.telegrams.put_nowait(telegram)
-            return None
+            return
         if isinstance(telegram.destination_address, IndividualAddress):
             if telegram.destination_address != self.xknx.current_address:
-                return None
-        return self.xknx.management.process(telegram)
+                return
+        self.xknx.management.process(telegram)
 
     async def send_telegram(self, telegram: Telegram) -> None:
         """Send telegram to connected device (either Tunneling or Routing)."""
@@ -486,13 +486,9 @@ class KNXIPInterfaceThreaded(KNXIPInterface):
         loop_loaded.set()
         self._thread_loop.run_forever()
 
-    async def _await_from_different_thread(
-        self,
-        coro: Awaitable[T],
-        target_loop: asyncio.AbstractEventLoop,
-    ) -> T:
+    async def _await_from_connection_thread(self, coro: Awaitable[T]) -> T:
         """Await coroutine in different thread."""
-        fut = asyncio.run_coroutine_threadsafe(coro, target_loop)
+        fut = asyncio.run_coroutine_threadsafe(coro, self._thread_loop)
         finished = threading.Event()
 
         def fut_finished_cb(_: concurrent.futures.Future[T]) -> None:
@@ -500,13 +496,13 @@ class KNXIPInterfaceThreaded(KNXIPInterface):
             finished.set()
 
         fut.add_done_callback(fut_finished_cb)
-        # wait on that event in an executor, yielding control to current loop
-        await asyncio.get_running_loop().run_in_executor(None, finished.wait)
+        # wait on that event in an executor, yielding control to _main_loop
+        await self._main_loop.run_in_executor(None, finished.wait)
         return fut.result()
 
     async def start(self) -> None:
         """Start KNX/IP interface."""
-        return await self._await_from_different_thread(self._start(), self._thread_loop)
+        return await self._await_from_connection_thread(self._start())
 
     async def stop(self) -> None:
         """
@@ -515,26 +511,22 @@ class KNXIPInterfaceThreaded(KNXIPInterface):
         Can not be restarted, create a new instance instead.
         """
         if self._interface is not None:
-            await self._await_from_different_thread(
-                self._interface.disconnect(), self._thread_loop
-            )
+            await self._await_from_connection_thread(self._interface.disconnect())
             self._interface = None
         self._thread_loop.call_soon_threadsafe(self._thread_loop.stop)
         self.connection_thread.join()
 
-    async def telegram_received(self, telegram: Telegram) -> list[Telegram] | None:
+    def telegram_received(self, telegram: Telegram) -> None:
         """Put received telegram into queue. Callback for having received telegram."""
-        return await self._await_from_different_thread(
-            super().telegram_received(telegram), self._main_loop
-        )
+        self._main_loop.call_soon_threadsafe(super().telegram_received, telegram)
 
     async def send_telegram(self, telegram: Telegram) -> None:
         """Send telegram to connected device (either Tunneling or Routing)."""
         if self._interface is None:
             raise CommunicationError("KNX/IP interface not connected")
 
-        return await self._await_from_different_thread(
-            self._interface.send_telegram(telegram), self._thread_loop
+        return await self._await_from_connection_thread(
+            self._interface.send_telegram(telegram)
         )
 
     async def gateway_info(self) -> GatewayDescriptor | None:
@@ -542,7 +534,7 @@ class KNXIPInterfaceThreaded(KNXIPInterface):
         if self._gateway_info is not None:
             return self._gateway_info
         if isinstance(self._interface, _Tunnel):
-            return await self._await_from_different_thread(
-                self._interface.request_description(), self._thread_loop
+            return await self._await_from_connection_thread(
+                self._interface.request_description()
             )
         return None