+++ docs/changelog.md
@@ -12,6 +12,7 @@ nav_order: 2
 
 - Parse T_Data_Broadcast TPCI. Forward these telegrams to the Management class.
 - KNXIPHeader total_length is 2 bytes long. There are no reserved bytes.
+- Revert handling L_Data.req frames for incoming device management requests.
 
 # 2.2.0 Expose cooldown 2022-12-27